Get rid of useCallback and move pure functions to utils file
